Centralize data lake file handling (Python) (#471)
