+++ source/databricks/tests/datamigration/test_uncommitted_migrations_count.py
@@ -18,8 +18,6 @@ from unittest.mock import patch
 
 from package.datamigration.uncommitted_migrations_count import (
     _get_valid_args_or_throw,
-    _get_file_system_client,
-    _download_file,
     _download_committed_migrations,
     _get_all_migrations,
     _get_uncommitted_migrations_count,
@@ -54,36 +52,17 @@ def test__uncommitted_migrations_count__when_invoked_with_correct_parameters__su
     _get_valid_args_or_throw(command_line_args)
 
 
-@patch("package.datamigration.uncommitted_migrations_count.DataLakeServiceClient")
-def test__get_file_system_client__calls_service_client_with_container_name(
-    mock_data_lake_service_client,
-):
-
-    # Arrange
-    dummy_storage_account_name = "my_storage"
-    dummy_storage_key = "my_storage"
-    dummy_container_name = "my_container"
-
-    # Act
-    _get_file_system_client(
-        dummy_storage_account_name, dummy_storage_key, dummy_container_name
-    )
-
-    # Assert
-    mock_data_lake_service_client.return_value.get_file_system_client.assert_called_once_with(
-        dummy_container_name
-    )
-
-
-@patch("package.datamigration.uncommitted_migrations_count._download_file")
+@patch("package.datamigration.uncommitted_migrations_count.download_csv")
 def test__download_committed_migrations__returns_correct_items(
-    mock_download_file,
+    mock_download_csv,
 ):
     # Arrange
     migration_name_1 = "my_migration1"
     migration_name_2 = "my_migration2"
-    csv_string = f"{migration_name_1}\r\n{migration_name_2}\r\n"
-    mock_download_file.return_value = str.encode(csv_string)
+    mock_download_csv.return_value = [
+        [migration_name_1],
+        [migration_name_2],
+    ]
 
     # Act
     migrations = _download_committed_migrations("", "", "")
@@ -92,12 +71,12 @@ def test__download_committed_migrations__returns_correct_items(
     assert migrations[0] == migration_name_1 and migrations[1] == migration_name_2
 
 
-@patch("package.datamigration.uncommitted_migrations_count._download_file")
+@patch("package.datamigration.uncommitted_migrations_count.download_csv")
 def test__download_committed_migrations__when_empty_file__returns_empty_list(
-    mock_download_file,
+    mock_download_csv,
 ):
     # Arrange
-    mock_download_file.return_value = b""
+    mock_download_csv.return_value = []
 
     # Act
     migrations = _download_committed_migrations("", "", "")
@@ -110,7 +89,7 @@ def test__download_committed_migrations__when_empty_file__returns_empty_list(
 @patch(
     "package.datamigration.uncommitted_migrations_count._download_committed_migrations"
 )
-def test__get_uncommitted_migrations_count__when_no_migration_needed__returns_zero(
+def test__get_uncommitted_migrations_count__when_no_migration_needed__returns_0(
     mock_download_committed_migrations, mock_get_all_migrations
 ):
     # Arrange
@@ -133,7 +112,7 @@ def test__get_uncommitted_migrations_count__when_no_migration_needed__returns_ze
 @patch(
     "package.datamigration.uncommitted_migrations_count._download_committed_migrations"
 )
-def test__get_uncommitted_migrations_count__when_one_migration_needed__returns_one(
+def test__get_uncommitted_migrations_count__when_one_migration_needed__returns_1(
     mock_download_committed_migrations, mock_get_all_migrations
 ):
     # Arrange
